plugins/CatCat/requirements.txt
PyYAML
aiohttp
ncatbot
aiofiles